requirements.txt
accelerate==0.34.2
diffusers==0.30.2
duckdb==1.2.0
matplotlib==3.10.3
nltk==3.9.1
numpy==2.2.6
pandas==2.2.3
Pillow==11.2.1
pyarrow==14.0.2
PyYAML==6.0.2
segment_anything==1.0
setuptools==65.5.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.66.5
transformers==4.49.0
Unidecode==1.3.8
wandb==0.16.5
